ClearVolumeClient now is smart enough to change the byte depth and number of layers of a renderer on the fly

+++ src/java/clearvolume/volume/Volume.java
@@ -158,7 +158,7 @@ public class Volume<T> implements ClearVolumeCloseable
 		return mDimensionsInVoxels;
 	}
 
-	public long getBytesPerVoxel()
+	public int getBytesPerVoxel()
 	{
 		if (mType == Byte.class || mType == byte.class)
 		{
